[build] remove disabled option from go_test template
For the few uses of this option, use a tag in the environment instead.
This results in the tests being included in tests.json, but not being
run in infrastructure. Before they were absent from tests.json.
This means the tests are now runnable by `fx test`.

Having only one way of doing something simplifies the code.

Bug: 55208
Change-Id: I7129652a0dddd701f64c9eb4ce42bef4791c1247
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/420923
Reviewed-by: David Turner <digit@google.com>
Reviewed-by: Erick Tryzelaar <etryzelaar@google.com>
Commit-Queue: Gary Miguel <garymm@google.com>