[rust] Make handle invalid() a const fn
This can be called in const contexts e.g. for initializing a
struct with invalid fields or for performing branches based on
invalid / not.

In these cases, the drop function won't be called (which would
be a no-op anyways if it were).

This is not expected to have much impact on performance today.

Change-Id: I23d581bf1e5127b94e6255e272637d3d95939a36
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/423400
Reviewed-by: Mitchell Kember <mkember@google.com>
Reviewed-by: Pascal Perez <pascallouis@google.com>
Testability-Review: Mitchell Kember <mkember@google.com>
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>