[fint] Check in static.proto
This protobuf contains all current fuchsia.proto spec fields that are
needed by the build, and thus will eventually be consumed by fint.

I'm also checking in the Go generated code for the protobuf. I spent a
while trying to add a GN target to ensure that the generated code is
always up to date with the protobuf, but I wasn't able to get that
working, so for now we'll just have to remember to manually re-run `go
generate` after updating the proto.

Bug: 59040
Change-Id: Iece70f00e61995a0cdc269669945a3120d7b3630
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/421555
Testability-Review: Oliver Newman <olivernewman@google.com>
Commit-Queue: Oliver Newman <olivernewman@google.com>
Reviewed-by: George Kulakowski <kulakowski@google.com>
Reviewed-by: Gary Miguel <garymm@google.com>