[zbitl] Introduce storage test traits
This change simplifies the shape of the storage-specific helpers
required for the general test logic, especially in a way that will make
testing the introduction of copy/cloning storage traits easier.

We replace the state-carrying "Io" structs with simpler trait structs
of static methods. These are moved into headers in anticipation of
tests for copy traits (forthcoming), at which point we'll want to test
the cross-product of the storage types and have all of the
specializations in a single translation unit.

Also, we fix a bug in which the stdio tests were not closing their file
pointers.

Bug: 57244
Change-Id: Iabac7d29bec457be11799d67a55b61c33cfb6363
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/422094
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Testability-Review: Roland McGrath <mcgrathr@google.com>