[storage] Fix rw-workers-tests.
rw-workers-test was testing files backed in memfs rather than the
filesystem under test.  This was because it forgot to append a "/" to
the mount path.

To address this, I've changed it so that the mount path always includes
a trailing slash.

Change-Id: I4abf3599ac32f65a3bfd165a2ffbc1f0f3989a47
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/423336
Testability-Review: Chris Suter <csuter@google.com>
Commit-Queue: Chris Suter <csuter@google.com>
Reviewed-by: Martin Lindsay <mlindsay@google.com>