[fs-management] Recognise vbmeta partitions.
Recognise the magic header 'AVB0' of vbmeta partitions.

This is a work-around of them being misrecognised as an MBR partition if
the last two bytes of the last sector happen to be 0x55 0xAA, of which
there is a roughly 1 in 65536 chance. (These bytes tend to be ~random in
vbmeta partitions).

Bug: 59374
Change-Id: I86faac94bc43dc3eef99774ad31e835f968e00f3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/423340
Commit-Queue: David Greenaway <dgreenaway@google.com>
Testability-Review: David Greenaway <dgreenaway@google.com>
Reviewed-by: Chris Suter <csuter@google.com>