[flatland] Engine unittest harness.
The Flatland Engine will have two main inputs: the InstanceMap from
the UberStructSystem and the LinkTopologyMap from the LinkSystem.
Typically, these two are populated by Flatland instances, but this is
not desirable from a unittesting standpoint since it depends on a lot
of functionality from the Flatland implementation.

This harness includes the FakeFlatlandSession, which includes the
barebones necessary to create and manage the data associated with a
Flatland instance. Specifically, it provides a convenient way to set
up links and local topologies, but leaves the creation of images and
matrices to the tests themselves.

The file includes an example test that should be deleted when actual
Engine tests are being written.

TEST=fx test -o flatland_unittests
MULTIPLY: flatland_unittests (fuchsia): 30

Change-Id: Ibc72c13977ce74d1e9c7c02dd8c21d770cc437ab
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/422106
Commit-Queue: Evan Birenbaum <birenbaum@google.com>
Reviewed-by: Chris Fontas <cfontas@google.com>
Testability-Review: Chris Fontas <cfontas@google.com>