[thermal] Remove MMIO offset from driver
The aml-thermal-s905d2g/-legacy drivers include a baked in 0x800 offset
that is used to address into the provided MMIO region for controlling
the sensor. Because of this, the MMIO address that is provided by the
board file has 0x800 subtracted from the true sensor base address.

It is unlcear why the offset was ever used in the first place. In its
current form, it only serves to add confusion because the MMIO address
that must be provided to the sensor driver is not part of the SoC memory
map documentation.

Test: fx test aml-thermal-s905d2g-test aml-thermal-s905d2g-legacy-test
Test: boot up Astro and ensure power_manager is still receiving
  temperature samples
Change-Id: I917eac2457752865ed656d3712edef3ccc609929
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/423725
Reviewed-by: Gurjant Kalsi <gkalsi@google.com>
Reviewed-by: Braden Kell <bradenkell@google.com>
Reviewed-by: Ruchira Ravoori <ravoorir@google.com>
Testability-Review: Gurjant Kalsi <gkalsi@google.com>
Testability-Review: Braden Kell <bradenkell@google.com>
Commit-Queue: Patrick Shickel <pshickel@google.com>