[audio_core] Fix Gain::GetGainDb
There is additional logic in `GetGainScale` that was not being applied
in GetGainDb. Fix this by simply defining GetGainDb using GetGainScale.

This fixes an issue where we can report an incorrect stream gain/volume
to effects. This can occur because `GetGainDb` was referencing values
that are cached and only updated by `GetGainScale`. In the sampler,
there is a fast-path that checks `Gain::IsSilent`, which bypasses this
`GetGainScale` method (which is what allowed this to work before).

Test: fx test audio_core_unittests
Change-Id: Ib891e3f22c5e1e5bd75778145b3042a3b1709d5a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/423594
Reviewed-by: Tom Bergan <tombergan@google.com>
Testability-Review: Tom Bergan <tombergan@google.com>
Commit-Queue: Tim Detwiler <tjdetwiler@google.com>