[fx] Add options to fx get-build-artifacts
This change introduces three new flags to `fx get-build-artifacts`:
* --expect-one: will fail if zero or more than one artifact is found
* --os: filter by os
* --cpu: filter by archictecture

--expect-one will be useful in the upcoming deprecation of
image_paths.sh, in which we'll look up images and only expect to find
one; the additional filters will be useful when looking up tools of a
specific os+cpu on a remote machine, for use in the -remote scripts.

Also the modes of `flash` and `zbi-tests` are introduced.

Change-Id: Iee2871823cb3ded72f758d8b77396b5fbe1656b9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/424125
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Testability-Review: Roland McGrath <mcgrathr@google.com>