[build]Add fvm fastboot to build
The CL adds fastboot fvm format into the build. The rule for fvm.blk is
kept, but is not enabled by default. It will only be built when the
build is not configured to generate a fastboot fvm format or user runs
"fx netboot" which requires to build an fvm.blk.

Bug: 48803

Change-Id: I338e86adc9c0fe5c10c077edc2f2584d1a13fd10
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/421751
Commit-Queue: Yecheng Zhao <zyecheng@google.com>
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: David Pursell <dpursell@google.com>
Testability-Review: Joshua Seaton <joshuaseaton@google.com>