[thermal] Remove MMIO offset for trim info
The aml-thermal-s905d2g/-legacy drivers currently expect an MMIO region
corresponding to GPIO_A0_BASE, then use a fixed offset to access the
register containing the trim info (used to derive calibration info).

This works for the single driver instance. But it won't be possible to
reuse the driver for other thermal device instances. The reason is that
each thermal instance gets its trim info from a register at a different
offset from GPIO_AO_BASE. Therefore the driver cannot expect that the
offset will always be the same.

To resolve this, pass in the MMIO address of the trim register in
question and remove the offset from the driver.

Test: fx test aml-thermal-s905d2g-test aml-thermal-s905d2g-legacy-test
Test: boot up on Astro and verify temperature samples are received and
  within valid range
Change-Id: Ic7cada54aafb531a69dd3fefb0c77c189f5a4dd3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/424026
Reviewed-by: Gurjant Kalsi <gkalsi@google.com>
Testability-Review: Gurjant Kalsi <gkalsi@google.com>
Commit-Queue: Patrick Shickel <pshickel@google.com>