[build] use host_test in go_test
This has the effect of generating a wrapper script for all host-side go
tests. This is a necessary step towards removing support for args in
test_spec.gni, which in turn is necessary to resolve the attached bug.

Also remove a install_host_tools() declaration for size_checker_tests.
size_checker is a host tool, but its tests aren't, so this didn't really
make sense to begin with. But now that the go_test() target doesn't
generate the output file that install_host_tools expected it was causing
errors.

Various other minor clean-ups in go_test.

Bug: 55208
Change-Id: I3a0e63f651cbe815bb638ea498bc39f3c05c14cd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/420924
Reviewed-by: David Turner <digit@google.com>
Reviewed-by: Gordon Tsai <gtsai@google.com>
Reviewed-by: Marc-Antoine Ruel <maruel@google.com>
Commit-Queue: Gary Miguel <garymm@google.com>