[debugger] Don't recursively enqueue BP steps.
In some error cases we can try to step over the same breakpoint twice.
This can happen if a memory write fails as we try to remove the
breakpoint to step over it (see added comment in debugged_process.h for
more).

This patch prunes the queue so a thread can never be stuck waiting for
itself.

Change-Id: I7b0dfb8fdb1e7f28391ee7caff96442f946c3625
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/424657
Reviewed-by: Dangyi Liu <dangyi@google.com>
Testability-Review: Brett Wilson <brettw@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>