[build] add fastboot_flash image metadata
Seeds the image rules with fastboot_flash metadata that gets put in the
generated images.json. Nothing uses this metadata yet, but we're working
towards making images.json the source-of-truth for image handling, so
this can be used in the future to replace things like the auto-generated
flash.sh script.

Test: images.json has the fastboot_flash fields (Astro)
Change-Id: I395c3bae668d6250d64fda098b75a28ed04f5512
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/423724
Commit-Queue: David Pursell <dpursell@google.com>
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Mike Comfoltey <comfoltey@google.com>