[time]Use PushSource API on network_time_service.
This starts to use the time source created earlier, and uses this to add
more comprehesive unittests for maintain_utc (note there is quite a lot
of duplication across these tests - that will get better in the future).

Bug 59408 was a problem with the inspect object falling out of scope
when maintain_utc finishes. This CL fixes that by passing an Arc::Mutex
into the async function so it isn't lost. After diagnostic rework I'll
be using interior mutability and so the mutex won't be required.

Bug: 56857, 59408
Change-Id: I50a8896e382427685ec066f06cbc1875b9a2a99a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/422887
Testability-Review: Satsuki Ueno <satsukiu@google.com>
Commit-Queue: Jody Sankey <jsankey@google.com>
Reviewed-by: Satsuki Ueno <satsukiu@google.com>