[ddk] Remove deprecated teardown APIs.
Added DdkUnbind and ddk::Unbindable, which are the same as
DdkUnbindNew and ddk::UnbindableNew, so we can soft transition
to the new naming.

Since MockDevice implements the FullDevice, this needed to be
migrated to the new naming. It looks like I also missed an
UnbindableDeprecated reference in a spi test that is not needed.

BUG=37831

Change-Id: Id3182f25eb6b023c1f3b9c1f21b28bf61d9c1c86
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/425002
Reviewed-by: Ruchira Ravoori <ravoorir@google.com>
Testability-Review: Jocelyn Dang <jocelyndang@google.com>
Commit-Queue: Jocelyn Dang <jocelyndang@google.com>