[flatland] Have Flatland schedule session updates.
The FrameScheduler has separate APIs for "registering a Present",
which indicates that a client wants to Present in the future and
allocates a PresentId to represent it, and "scheduling a Present",
which tells the FrameScheduler when the client actually wants that
Present to be applied. This change introduces the second API to the
FlatlandPresenter, which must happen after the acquire fences, and
allows the first step to happen immediately.

TEST=fx test -o flatland_unittests
MULTIPLY: flatland_unittests (fuchsia): 30

Change-Id: I03b05ce8ead6e6614bc458d4cd21437b6f3730fe
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/425008
Commit-Queue: Evan Birenbaum <birenbaum@google.com>
Reviewed-by: Mikael Pessa <mikaelpessa@google.com>
Testability-Review: Mikael Pessa <mikaelpessa@google.com>