[bt][rfcomm] Implement initial MuxCommand handling
Introduces parsing support for Frames with MuxCommands. Currently, the
headers of the MuxCommands are parsed and validated. The payload itself
(which is specific to a command) is not parsed. These are marked as
TODOs.

Introduces Encodable,Decodable traits which generalizes the packet
serialization/deserialization process for all MuxCommand frames.

Bug: 58681
Test: Added to bt-rfcomm-tests
Change-Id: I2aab5d961dc49214a4b45bfed7576925f2a678c0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/421194
Commit-Queue: Ani Ramakrishnan <aniramakri@google.com>
Reviewed-by: Marie Janssen 💖 <jamuraa@google.com>
Testability-Review: Marie Janssen 💖 <jamuraa@google.com>