[audio][test] Add a framework for running tests based on golden files
This is based on the design here:
https://docs.google.com/document/d/1pOfqq0pniWUqEM-044fX-BEw1omDe_jtrmXkpde-zkc

To be used in a future CL.

The hermetic_golden_test_update_goldens.sh script can be used to create
the initial golden files for a new test, or to update the golden files
of an existing test. It opens the following audio buffers in Audacity
(or a user-specified program):

1. The test input
2. The output ring buffer
3. The portion of the output ring buffer that is expected to be
   non-silent
4. The expected output, which should match (3)

The script then asks the user whether they want to overwrite the
existing golden file (4) with the new golden (3).

The framework includes a waveform test (based on goldens) and an impulse
test (which doesn't need golden output files).

Bugs: 49807
Change-Id: I2d30f7bb1ee39822d4c7f33fee9afc3933da47b9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/393838
Commit-Queue: Tom Bergan <tombergan@google.com>
Reviewed-by: Tim Detwiler <tjdetwiler@google.com>
Testability-Review: Tim Detwiler <tjdetwiler@google.com>