[goldfish] Support composite device in goldfish_control driver.
This change adds support of goldfish-control-2 composite device to
"goldfish-control" driver so that it can access the goldfish address
space device.

Since only one control device is supported on the device (otherwise
there will be multiple goldfish heaps and goldfish display devices),
we do not create a new driver, instead we add some #ifdefs to the
original driver to control which device to bind to.

Currently whether to use the composite device is controlled by a
build-time argument "goldfish_control_use_composite_device" (default
to false) to control which type of device the driver binds to. This
will be switched to true by defualt in a subsequent change.

TEST=goldfish_test with goldfish_control_use_composite_driver = true
     goldfish_test with goldfish_control_use_composite_driver = false

Change-Id: I5411a432c65010cdaf64ca259fa5749ec2461dd4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/424654
Commit-Queue: Yilong Li <liyl@google.com>
Reviewed-by: David Reveman <reveman@google.com>
Testability-Review: David Reveman <reveman@google.com>