[fidlgen] Write unformatted generated code when formatting fails
Currently if formatting fails, we output an error but don't write to the
files being generated. This can make it hard to debug fidlgen changes
that result in formatting errors.

This CL changes the behavior so that a file is still written if
formatting fails, but it is written with unformatted text.

Test: success cases are tested through fidlgen tests. There are no tests
for failure cases, but tested manually.

Change-Id: Ie228c31edee78d8be0e6988fb3e89232255b718c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/423917
Reviewed-by: Pascal Perez <pascallouis@google.com>
Testability-Review: Pascal Perez <pascallouis@google.com>
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>