[traceutil] Resolve SSH credentials from manifest
The location of the default SSH credentials used in connecting to a
Fuchsia device has changed. This change updates traceutil to use a
manifest to find the appropriate path to the default SSH private key
(identity) and authorized_keys.

The aforementioned manifest is generated by a jiri hook on update and is
not expected to be absent in any Fuchsia tree unless the user has
removed it.

Test: `fx traceutil record` records traces (including WFH setup)

Change-Id: I7ba09f76677aca5d62c54ca36d7e64be19fbb558
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/419480
Reviewed-by: Nathan Rogers <nathanrogers@google.com>
Testability-Review: Nathan Rogers <nathanrogers@google.com>
Commit-Queue: Wesley Moy <wesleymoy@google.com>