[build][alln] Add kernel.phys32 Zircon toolchain suite.
Add BUILD.gn files to support building the kernel.phys32 loader
with the Fuchsia build. This is only used to verify that the build
commmands are identical to the Zircon build ones, thanks to the
compare_toolchains.py script. This doesn't build anything at the
moment.

This introduces a few BUILD.gn that mimic a BUILD.zircon.gn file,
generally in the same directory. These are not symlinks because
only a tiny part of the zircon build file is replicated for this
CL.

Bug: 54160
Testing: RUnnint the script [1] should succeed.

[1] zircon/public/canaries/compare_toolchains.py --clean /tmp/results

Change-Id: I49c38e640e3db540261445552e715948c0117051
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/425315
Testability-Review: David Turner <digit@google.com>
Reviewed-by: Shai Barack <shayba@google.com>
Reviewed-by: Petr Hosek <phosek@google.com>
Reviewed-by: P.Y. Laligand <pylaligand@google.com>
Commit-Queue: David Turner <digit@google.com>