[fidlc] Parse modifiers in any order
This CL enables fidlc to parse the modifiers "strict", "flexible", and
"resource" in any order. It improves error messages when they are used
on declarations that do not support them, when provided multiple times,
or when conflicting (as in "strict flexible"). These errors are also now
recoverable: parsing continues as if the offending modifier was omitted.

Fixed: 59467
Test: fx test fidl-compiler
Change-Id: Ided089fb896b14a0deea1586c0b76bbeabbce174
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/425897
Commit-Queue: Mitchell Kember <mkember@google.com>
Reviewed-by: Pascal Perez <pascallouis@google.com>
Testability-Review: Pascal Perez <pascallouis@google.com>