[build][fuzzing] Split fuzzer instrumentation and linking
This CL makes the "fuzzer" variants more flexible in order to support
building multi-process fuzzers in which the engine is linked against
libFuzzer, but the code under test is only instrumented. It modifies the
original variant to no longer link libFuzzer, provides a config that
links the static compiler runtime library, and modifies the `fuzzer`
template to include it.

These changes have no effect on existing fuzzers, but they allow the
creation of new, multi-process fuzzers that have the fuzzing engine and
instrumented code under test in separate processes.

Bug: 45347
Change-Id: I87b172f22a362b6632d49e26a37155611b654e1d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/421176
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Reviewed-by: Matthew Dempsky <mdempsky@google.com>
Commit-Queue: Aaron Green <aarongreen@google.com>