[rust] Update paste crate to 1.0
rust_icu (which is our own crate) was also updated to 0.3.1, which bumps
its own dependency on paste to 1.0 so that we only need to vendor a
single copy of paste.

ecma402_traits is a "new" crate but is from the same google/rust_icu
repository.

Change-Id: I18eb48ca48d945342e5f9c3189146d2705045178
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/423703
Commit-Queue: Bryan Henry <bryanhenry@google.com>
Testability-Review: Filip Filmar <fmil@google.com>
Reviewed-by: Dan Johnson <computerdruid@google.com>
Reviewed-by: Filip Filmar <fmil@google.com>