[debugger] Remove error log for Limbo.
This error writes to the system log whenever limbo isn't enabled which
is the normal state for most developers. It was causing confusion.

The only behavior change is that on error waiting for limbo, the hanging
get for limbo is no longer retried on error.

Minor stylistic updates to the limbo file.

Bug: 58403
Change-Id: Iff46cc1d5e5b2e2d20cfcc316c91fd04917c48fb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/426017
Reviewed-by: Dangyi Liu <dangyi@google.com>
Testability-Review: Brett Wilson <brettw@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>