[affectedtests] Take an axe to affectedtests/testdata
People were complaining that the checked-in tests.json was hitting their
greps, so we fixed the glitch. :p

Delete most of the testdata except the entries that we actually exercise
in the test, plus a few duds for shits & giggles.

Change-Id: Ib8ff3863230adb8bbbb4835e57a7c3eda31755b3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/426135
Testability-Review: Shai Barack <shayba@google.com>
Reviewed-by: Jay Zhuang <jayzhuang@google.com>
Commit-Queue: Shai Barack <shayba@google.com>