[console] Output syslog to serial
This change allows console to act as a LogListener and output the syslog
to the serial console.

This solves the problem of migrating to syslog, but not having access to
logs from the serial console without explicitly running log_listener.

Fixed: 59371
Change-Id: Ie532c9c9257058ff394b8cc704d931acfde0d4c5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/423854
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: Adam Perry <adamperry@google.com>
Commit-Queue: Abdulla Kamar <abdulla@google.com>
Testability-Review: Adam Perry <adamperry@google.com>