[end_to_end] Remove audio_record_test
audio_record_test is currently configured as part of terminal and to run
on astro only. Since the test communicates directly with audio drivers,
it requires real hardware. However, we no longer have any core.astro or
terminal.astro CQ bots, so this test has not been running in CQ.

I tried adding sherlock_env and nuc_env, which enables this test on the
core.sherlock-debug and core.fuchsia-x86-* CQ bots. This revealed a
problem in the test: on sherlock, there are actually two built-in audio
devices, one of which is actually a BT device use for testing only. It
turns out this device gets selected by default because it's device #0.
The test does not work with this BT device -- it must use the actual
hardware audio device.

Making this test work on any device will be non-trivial: we need to call
`audio-driver-ctl info`, parse the result, locate the correct device
number by unique ID, parse that device's local device id (`n`), then
call `audio-driver-ctl record -d n`. While this is doable, it makes the
test depend on the text output of audio-driver-ctl, which is not
desirable.

This test was originally added by fxr/279256 as a regression test to
ensure that the microphone records something. It's not indending to test
`audio-driver-ctl`, just the microphone. It turns out we have another
test with the same purpose: audio_core_hardware_test.

Hence, since it's not trival to fix audio_record_test in a clean way,
and since the test is redundant with another test, we have decided to
remove audio_record_test entirely.

While working on this bug, I found a small error handling bug in
audio-device-stream.cc, which has also been fixed by this CL.

Change-Id: Id9a73f3f8611ba0f2857c72a020000028ddf8d7a
Fixed: 58634
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/422111
Reviewed-by: James Tucker <raggi@google.com>
Reviewed-by: Andres Oportus <andresoportus@google.com>
Commit-Queue: Tom Bergan <tombergan@google.com>
Testability-Review: Andres Oportus <andresoportus@google.com>