[elflib] List sources from a given binary
Useful for debuginfod-related efforts, but also for a simple utility I
have in mind for dumping the sources for a given binary (`srcdump`?).
In particular, it would be nice to occasionally use the latter to audit
what's in the kernel or, eventually, physboot.

Bug: 49702
Change-Id: I7d7415766086d465d54d2a23d25cb872cece5341
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/405939
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Reviewed-by: Anthony Fandrianto <atyfto@google.com>
Testability-Review: Roland McGrath <mcgrathr@google.com>