[kernel] Mutex: Assert no spinlocks are held when acquiring a Mutex
When attempting to acquire a mutex, a thread should not hold any
spinlocks; if it does, it risks blocking while a spinlock is held,
which is not allowed by the Fuchsia locking rules.

We already have an assertion if we attempt to block while any
spinlocks are held, but that assertion will not catch uncontended
Mutex acquisitions. Add an assert very early in Mutex::Acquire to
catch this case.

Change-Id: Ic2de0283b9a7ad89846b6d293e1a6bda5d7c6b7c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/300275
Commit-Queue: Venkatesh Srinivas <venkateshs@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Testability-Review: Adrian Danis <adanis@google.com>
Testability-Review: Nick Maniscalco <maniscalco@google.com>