[blobfs][reland] Disable cache eviction
Reland of https://fuchsia-review.googlesource.com/c/fuchsia/+/421454.

Once a blob is first opened, it will never be evicted, even after the
last handle to the blob closes.

This improves the performance of later accesses to a file, especially in
degenerate cases where a file is repeatedly reopened (which happens
during testing, particularly for test runners that spawn a new process
for each test case for hermeticity).

For example, improves the runtime of blobfs-unittest (repackaged as a V2
test component) from 4m20s to 50s, a 5x reduction in runtime.
(Subsequent local re-runs are even faster, around 20-30s).

Memory is reclaimed by the system as needed by the kernel page eviction
mechanism. Data pages that have not been recently accessed are reclaimed
when the system is under memory pressure. The only additional memory
overhead that is not reclaimed is blobfs' internal memory use for
metadata (particularly, merkle trees).

Bug: 58415
Bug: 59601
Change-Id: I884ed5281aba6ffe80ef713c829c8c1336f80b65
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/425207
Commit-Queue: James Sullivan <jfsulliv@google.com>
Reviewed-by: Adam Barth <abarth@google.com>