[nix] Add skeleton epoll implementation
This CL creates a new library, nix, for incubating new POSIX-compatible
APIs. The first API to be incubated in this library is epoll.

The implementation of epoll in this CL is just stubs that return
failure. This purpose of the CL is to create the structure (e.g.,
library, headers, target and host tests) needed to build out the
implementation.

This CL is the start of a "revert" of
09fe3b522ce70e25655c34bed039d2ee88796b07, which removed our original
epoll implementation. The CL also draws upon elements from
https://fuchsia-review.googlesource.com/c/fuchsia/+/364058, which will
also be used more in future CLs that add real functionality.

Test: nix-test, nix_hosttests
Change-Id: If6113999276fdbc17e54aecdf1b20606b9cec80f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/426414
Reviewed-by: George Kulakowski <kulakowski@google.com>
Reviewed-by: Steve Austin <steveaustin@google.com>
Testability-Review: George Kulakowski <kulakowski@google.com>
Commit-Queue: Adam Barth <abarth@google.com>