[blobfs] Convert blobfs tests to packages.
Remove the "short" integration test. This test triggers some
high-severity logs that result in it being marked as failing despite
passing otherwise. Rather than tracking these down, this test is removed
along the lines of a TODO written by rvargas in the test code that it's
not actually testing something useful and should probably be removed.

Add a flag to the fs-test-support library to control whether the test
is in a package or not. Package tests require an additional path be
injected. The default is packaged since this is the direction we're
going, but the existing minfs and tests for the fs-test-support itself
all required opting-out of this feature to continue running.

Document the "all_source" config which I found confusing when updating
the build.

Bug: 52911

Change-Id: I7375565de10b587e3535afc397c09687156f3194
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/421736
Commit-Queue: P.Y. Laligand <pylaligand@google.com>
Reviewed-by: Bryan Henry <bryanhenry@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: Gianfranco Valentino <gevalentino@google.com>
Testability-Review: Gianfranco Valentino <gevalentino@google.com>