[components] Speculative fix for Hub v2
The stop event could be fired after destruction which would cause the
hub to crash component manager. This CL ensures that the stop event
happens before destruction.

This also fixes an underlying shutdown bug: stop and shutdown must be
serialized with respect to one another so that we cannot stop and
restart a component during shutdown. The shutdown bit must be set to
indicate that a component has stopped and cannot be started again. Prior
to this change, it is possible for that component to not be set during
shutdown due to a race with a stop.

This CL makes Stop an Action and adds a special case to make sure that
stop and shutdown actions serialize with one another.

Bug: 58378
Change-Id: I8ccfa1deed368f2ccb77cde0d713f3af221f7450
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/425654
Commit-Queue: Fady Samuel <fsamuel@google.com>
Reviewed-by: Gary Bressler <geb@google.com>
Testability-Review: Gary Bressler <geb@google.com>