[fuchsia-hyper] Fix panics, add new errors
In http://fxbug.dev/59391, the netstack crashed due to minfs filling up.
This caused the package resolver to panic when it tried to communicate
with fuchsia.posix.socket.Provider in order to resolve an interface name
to a scope id.

This patch switches from panicking to returning an error when it cannot
communicate with fuchsia.posix.socket.Provider. In addition, this changes
fuchsia-hyper to error out if the interface index is larger than 32
bits.

Test: Added tests for these cases.
Fixed: 59384
Change-Id: I47224872f2ba803e62051c734126b98e891eb317
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/423695
Reviewed-by: Tamir Duberstein <tamird@google.com>
Reviewed-by: Dan Johnson <computerdruid@google.com>
Commit-Queue: Erick Tryzelaar <etryzelaar@google.com>
Testability-Review: Tamir Duberstein <tamird@google.com>