[kernel][vm] Test for merging markers
Adds a unittest to validate that the migrate_fn callback of
VmPageList::MergeFrom is not called for markers. This is the current
behavior, but in the implementation and interface description, but it
is easy to violate when modifying the page list implementation and
there are no other tests for it.

Change-Id: I150dc2c9b9b0b5d57b82f56b2a15e69f0a617ed6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/425556
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Testability-Review: Rasha Eqbal <rashaeqbal@google.com>
Commit-Queue: Adrian Danis <adanis@google.com>