[docs][tests] Split the "Running tests as components" into concepts and how-to guides.
This change only splits the existing "Running tests as components" guide
into a concept doc and a how-to guide.

Further editing the content of these docs will take place once this
change is submitted -- this allows the updates of content to be viewed,
compared and commented on Gerrit.

Change-Id: I97a6aa276af91ede457d989ddde867a9d1434b37
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/426023
Commit-Queue: Kyo Lee <kyol@google.com>
Reviewed-by: Isma Cortes <isma@google.com>