[a11y] Implements recover A11Y Focus action.
This is a simple action to be executed when a semantic tree is updated. The A11Y focus will be checked to see if it still exists. If not, it is cleared, and as a result, so is the highlight.

This action is expected to be improved as we learn what are good ways of recovering the focus.

Change-Id: I255a32cf5dea360613c2fb1fb01a817b75b78b68
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/427015
Commit-Queue: Lucas Radaelli <lucasradaelli@google.com>
Reviewed-by: Alexander Brusher <abrusher@google.com>
Reviewed-by: Tess Eisenberger <teisenbe@google.com>
Testability-Review: Tess Eisenberger <teisenbe@google.com>