[zxcrypt] Allocate slices from 0, not 1
When running atop FVM, zxcrypt needs to ensure that it has enough space
allocated for both its own superblock and backup superblock, but also
it needs to ensure that the zeroth block of the FVM that it exposes is
also preallocated, since that's the usual behavior of the FVM itself.

An error in calculation caused us to preallocate an additional slice
here which zxcrypt never used and which the volume sitting atop it might
also never use.

Tests: added new test verifying that when presented with a new volume
(0th slice allocated, all others unallocated), we allocate exactly one
additional slice in `Init()`, and that if we already have two slices
allocated, we do not attempt to allocate additional slices.

Change-Id: Id7cf638e5739173951d19894c668063e9db4651f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/427257
Reviewed-by: Aaron Green <aarongreen@google.com>
Testability-Review: Aaron Green <aarongreen@google.com>
Testability-Review: Drew Fisher <zarvox@google.com>
Commit-Queue: Drew Fisher <zarvox@google.com>