[build] disallow fuchsia tests to set args in test spec
The component manifest already has a mechanism for specifying args which
get passed to the component itself.
The args feature was only being used by prebuilt_test_manifest to change
the runner which runs the component, but we don't want to allow that,
since we rely on the runners to enforce very important policies on
tests (e.g. timeouts, log severities).
This change closes a loop-hole that would allow tests to subvert these
policies, and it also simplifies the build system.

A later change will remove support for the custom command from the
testrunner.

Bug: 58561
Change-Id: Iadcc6e36458362239f907279ffc3015fe39e5868
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/418483
Commit-Queue: Gary Miguel <garymm@google.com>
Reviewed-by: David Turner <digit@google.com>